devtools::test() pass: correct issues in access datasets in data/
running test in console works, but test() didn't, because calculation table.rda "cannot open connection". but turns out to access datasets in data/, can just call it (no .rda at the end) (b/c lazyloading).--no need for system.file or load("data/...")--but then because filename is a variable (vector), you need to use get(filename that's a variable) to access the file. NOTE: no .rda at the end for that variable.